Disable DSP performance voting
The HVX service will vote to boost the performance of the DSP when the
service is first used but never votes to slow down the DSP afterwards.
This may be a waste of power for the DSP when it is not being used.

This CL removes the performance voting. Qualcomm is sending newer
binaries on 5/29 with smarter performance voting control for the NN
service.

Bug: 74643806
Bug: 78249699
Test: mma
Change-Id: Iec0beacb8fc267ae50f5c2f5f881ce1d0d74f8f9